refactor with a common internal header
